requirements.txt
starlette
fastapi
pydantic
uvicorn
pytz
phpserialize
mysql-connector